[#76] refactor: 함수 및 상수 분리 리팩토링
