+++ src/constants/url.ts
@@ -1,0 +1,3 @@
+const SERVER_URL =
+  'http://ec2-43-203-87-227.ap-northeast-2.compute.amazonaws.com';
+export default SERVER_URL;
