+++ src/hooks/graph/transformGraphInfo.ts
@@ -6,16 +6,20 @@ import {
 } from '@/types/graph/GraphProps';
 
 export const transformResults = (results: SearchResult[]): GraphNode[] => {
-  return results.map(result => {return {
-    id: result.documentId.toString(),
-    title: result.title,
-    group: result.group,
-  }});
+  return results.map(result => {
+    return {
+      id: result.documentId.toString(),
+      title: result.title,
+      group: result.group,
+    };
+  });
 };
 
 export const transformLinks = (serverLinks: ServerLink[]): GraphLink[] => {
-  return serverLinks.map(link => {return {
-    source: link.parentDocumentId.toString(),
-    target: link.childDocumentId.toString(),
-  }});
+  return serverLinks.map(link => {
+    return {
+      source: link.parentDocumentId.toString(),
+      target: link.childDocumentId.toString(),
+    };
+  });
 };
